+++ TCRT/TCRT.cpp
@@ -3,7 +3,7 @@
 TCRT *TCRT::sensors[SENSOR_AMOUNT] = {NULL}; // static member declaration 
 int TCRT::sensorCount = 0;                   // static member declaration 
 
-TCRT::TCRT(PinName Pin, PinName DarlingPin, float v) : sensorPin(Pin), DarlingtonPin(DarlingPin), VDD(v), rIndex(0), senseNorm(0)
+TCRT::TCRT(PinName Pin, PinName DarlingPin, double v) : sensorPin(Pin), DarlingtonPin(DarlingPin), VDD(v), rIndex(0), senseNorm(0)
 {
     senseNormRolled[0] = 
     senseNormRolled[1] = 
@@ -40,7 +40,7 @@ void TCRT::pollSensors(void)
     };
 };
 
-float TCRT::getSensorVoltage(bool Volt)
+double TCRT::getSensorVoltage(bool Volt)
 {
     return (Volt ? (senseNorm) * VDD : sensorPin.read());
 };
